added in two get handlers, moved local apis to localapi.
