+++ package.json
@@ -36,12 +36,12 @@
   "dependencies": {
     "axios": "^0.16.1",
     "bcrypt": "^0.8.7",
-    "body-parser": "^1.17.1",
+    "body-parser": "^1.17.2",
     "create-react-app": "1.2.1",
     "dotenv": "^2.0.0",
-    "express": "^4.14.0",
+    "express": "^4.15.3",
     "jwt-simple": "^0.5.1",
-    "mongoose": "^4.6.8",
+    "mongoose": "^4.10.4",
     "passport": "^0.3.2",
     "passport-jwt": "^2.2.1",
     "passport-local": "^1.0.0",
